Из Program в методы расширения вынесены настройки для WebApplicationBuilder и WebAplication. Методы класса ServiceExtensions разнесены по соответсвующим сборкам.
