+++ AdvertisingPlatforms/ServiceCollection/ServiceExtensions.cs
@@ -1,47 +1,10 @@
-﻿using AdvertisingPlatforms.DAL.Repositories;
-using AdvertisingPlatforms.Domain.Interfaces.Services;
-using AdvertisingPlatforms.Domain.Interfaces.Services.FileHandling;
-using AdvertisingPlatforms.Domain.Models;
-using AdvertisingPlatforms.Business.Services.AdvertisingServices;
-using AdvertisingPlatforms.Business.Services.FileHandlingServices;
-using AdvertisingPlatforms.DAL.FileAccess;
-using AdvertisingPlatforms.DAL.Interfaces;
-using AdvertisingPlatforms.DAL.Repositories.Base;
-
-namespace AdvertisingPlatforms.ServiceCollection
+﻿namespace AdvertisingPlatforms.ServiceCollection
 {
+    /// <summary>
+    /// Extensions for service collection.
+    /// </summary>
     public static class ServiceExtensions
     {
-        /// <summary>
-        /// Method for registration advertising services
-        /// </summary>
-        public static void AddAdvertisingServices(this IServiceCollection services)
-        {
-            services.AddScoped<IAdvertisingPlatformsService, AdvertisingPlatformsService>();
-            services.AddScoped<ILocationsService, LocationsService>();
-        }
-
-        /// <summary>
-        /// Method for registration repository services
-        /// </summary>
-        public static void AddRepositoryServices(this IServiceCollection services)
-        {
-            services.AddScoped<IRepositoryReader, RepositoryReader>();
-            services.AddScoped<IRepositoryWriter, RepositoryWriter>();
-            services.AddScoped<Repository<Location>, LocationsFileRepository>();
-            services.AddScoped<Repository<AdvertisingPlatform>, AdvertisingPlatformsFileRepository>();
-        }
-
-        /// <summary>
-        /// Method for registration file services
-        /// </summary>
-        public static void AddFileServices(this IServiceCollection services)
-        {
-            services.AddScoped<IFileReader, FileReader>();
-            services.AddScoped<IFileParser, FileParser>();
-            services.AddScoped<IFileValidator, FileValidator>();
-        }
-
         /// <summary>
         /// Method for registration file services
         /// </summary>
@@ -51,11 +14,11 @@ namespace AdvertisingPlatforms.ServiceCollection
 
             services.AddSwaggerGen(options =>
             {
-                var xmlFile = $@"AdvertisingPlatforms.xml";
+                var xmlFile = @"AdvertisingPlatforms.xml";
                 var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
                 options.IncludeXmlComments(xmlPath);
 
-                var xmlFileDomain = $@"AdvertisingPlatforms.Domain.xml";
+                var xmlFileDomain = @"AdvertisingPlatforms.Domain.xml";
                 var xmlPathDomain = Path.Combine(AppContext.BaseDirectory, xmlFileDomain);
                 options.IncludeXmlComments(xmlPathDomain);
             });
